+++ src/main/java/car/domain/Game.java
@@ -1,4 +1,4 @@
-package car;
+package car.domain;
 
 import java.util.ArrayList;
 import java.util.List;
