+++ src/main/java/car/domain/RandomMoveChance.java
@@ -1,4 +1,4 @@
-package car;
+package car.domain;
 
 import java.util.Random;
