+++ src/test/java/car/domain/GameTest.java
@@ -1,4 +1,4 @@
-package car;
+package car.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
